requirements.txt
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.3
webdriver-manager==4.0.1
selenium==4.22.0
allure-python-commons~=2.13.5
requests

